building-flask-app/requirements.txt
altgraph==0.16.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
idna==2.8
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.1.1
modulegraph==0.17
networkx==2.3
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
py2app==0.19
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
urllib3==1.25.6
virtualenv==16.7.6
Werkzeug==0.16.0
